requirements.txt
langchain==0.2.16
langchain-community==0.2.16
langchain-openai==0.1.23
faiss-cpu==1.8.0
openai==1.40.0
python-dotenv==1.0.1
pypdf==4.3.1
python-docx==1.1.2
tiktoken==0.7.0
flask==3.0.3
werkzeug==3.0.3
httpx==0.27.0
